+++ NamEcommerce/Domain/NamEcommerce.Domain.Services/Extensions/WarehouseExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using NamEcommerce.Domain.Entities.Inventory;
 using NamEcommerce.Domain.Shared.Dtos.Inventory;
+using NamEcommerce.Domain.Shared.Enums.Inventory;
 
 namespace NamEcommerce.Domain.Services.Extensions;
 
@@ -13,7 +14,7 @@ public static class WarehouseExtensions
             PhoneNumber = warehouse.PhoneNumber,
             Address = warehouse.Address,
             ManagerUserId = warehouse.ManagerUserId,
-            WarehouseType = (int) warehouse.WarehouseType,
+            WarehouseType =  warehouse.WarehouseType,
             WarehouseNameKey = $"Enums.{typeof(WarehouseType).FullName}.{warehouse.WarehouseType}",
             IsActive = warehouse.IsActive
         };
